Converted EuiSteps to TS
